hoàn thiện task API

+++ src/main/java/com/example/workerrental/repository/CustomerRepository.java
@@ -3,6 +3,6 @@ package com.example.workerrental.repository;
 import com.example.workerrental.repository.entity.CustomerEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CustomerRepository extends JpaRepository<CustomerEntity,Integer> {
+public interface CustomerRepository extends JpaRepository<CustomerEntity,Long> {
 
 }
