+++ src/main/java/com/example/workerrental/repository/TaskRepository.java
@@ -3,5 +3,5 @@ package com.example.workerrental.repository;
 import com.example.workerrental.repository.entity.TaskEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TaskRepository extends JpaRepository<TaskEntity, Integer> {
+public interface TaskRepository extends JpaRepository<TaskEntity, Long> {
 }
