+++ src/main/java/com/example/workerrental/service/TaskService.java
@@ -5,9 +5,9 @@ import com.example.workerrental.model.TaskDTO;
 import java.util.List;
 
 public interface TaskService {
-    TaskDTO getTaskById(Integer id);
+    TaskDTO getTaskById(Long id);
     List<TaskDTO> getAllTasks();
     TaskDTO createTask(TaskDTO taskDTO);
-    TaskDTO updateTask(Integer id, TaskDTO taskDTO);
-    Boolean deleteTask(Integer id);
+    TaskDTO updateTask(Long id, TaskDTO taskDTO);
+    Boolean deleteTask(Long id);
 }
